refactor(app): share civilian/diplomacy/production panel test hosts
Extract duplicated Game builders and panel shells so near-cap diplomacy,
civilian stay-split, and production suites densify without behavior change
(Refs #4013).
