+++ app/test/civilian_units_panel_part2_test.dart
@@ -7,7 +7,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:colonizethis_app/core/services/app_event_handler/app_event_handler_scope.dart';
 import 'package:colonizethis_app/providers/games_provider.dart';
 import 'package:colonizethis_app/features/game/widgets/chrome/ct_circular_locate_button.dart';
 import 'package:colonizethis_app/features/game/widgets/chrome/ct_danger_text_button.dart';
@@ -31,50 +30,6 @@ void main() {
     game = buildCivilianPanelTestGame();
     humanPlayerIdWithUnits = game.players.first.id;
   });
-
-  Widget buildPanel({
-    required Game game,
-    required String humanPlayerId,
-    Orders currentOrders = const Orders(),
-    Map<String, List<String>> availableWorkTargets = const {},
-    AppEventBus? bus,
-    bool explorerOnly = false,
-    bool builderOnly = false,
-    String? prospectShortcutTargetTileKey,
-    String? exploreShortcutTargetTileKey,
-    String? buildImprovementShortcutTargetTileKey,
-  }) {
-    final resolvedBus = bus ?? AppEventBus.create();
-    final navigatorKey = GlobalKey<NavigatorState>();
-    return ProviderScope(
-      overrides: [
-        availableWorkTargetIdsForUnitProvider.overrideWith(
-          (ref, unitId) => availableWorkTargets[unitId] ?? const [],
-        ),
-      ],
-      child: MaterialApp(
-        navigatorKey: navigatorKey,
-        home: Scaffold(
-          body: CivilianPanelBusDialogHost(
-            bus: resolvedBus,
-            navigatorKey: navigatorKey,
-            child: CivilianUnitsPanel(
-              game: game,
-              humanPlayerId: humanPlayerId,
-              currentOrders: currentOrders,
-              bus: resolvedBus,
-              explorerOnly: explorerOnly,
-              builderOnly: builderOnly,
-              prospectShortcutTargetTileKey: prospectShortcutTargetTileKey,
-              exploreShortcutTargetTileKey: exploreShortcutTargetTileKey,
-              buildImprovementShortcutTargetTileKey:
-                  buildImprovementShortcutTargetTileKey,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
 
   group('CivilianUnitsPanel', () {
     testWidgets(
@@ -117,7 +72,7 @@ void main() {
         bus.on<LocateMapTileEvent>().listen((e) => locateEvent = e);
 
         await tester.pumpWidget(
-          buildPanel(game: miniGame, humanPlayerId: human, bus: bus),
+          buildCivilianPanel(game: miniGame, humanPlayerId: human, bus: bus),
         );
         await tester.pumpAndSettle();
 
@@ -283,7 +238,7 @@ void main() {
         bus.on<LocateMapTileEvent>().listen((e) => locateEvent = e);
 
         await tester.pumpWidget(
-          buildPanel(
+          buildCivilianPanel(
             game: gameWithPending,
             humanPlayerId: human,
             currentOrders: orders,
@@ -336,7 +291,7 @@ void main() {
         if (availableWorkTargets.isEmpty) return;
 
         await tester.pumpWidget(
-          buildPanel(
+          buildCivilianPanel(
             game: game,
             humanPlayerId: humanPlayerIdWithUnits,
             bus: bus,
@@ -407,7 +362,7 @@ void main() {
           },
         );
         await tester.pumpWidget(
-          buildPanel(
+          buildCivilianPanel(
             bus: bus,
             game: game,
             humanPlayerId: humanPlayerIdWithUnits,
@@ -484,7 +439,7 @@ void main() {
           },
         );
         await tester.pumpWidget(
-          buildPanel(
+          buildCivilianPanel(
             bus: bus,
             game: game,
             humanPlayerId: humanPlayerIdWithUnits,
