+++ app/test/diplomacy_panel_test.dart
@@ -14,94 +14,6 @@ import 'support/diplomacy_panel_test_support.dart';
 import 'support/panel_test_fixtures.dart';
 import 'support/widget_test_assets.dart';
 
-Game _gameWithNoDiscoveredFactions() {
-  const ow = 'oldWorld';
-  final p1 = Province(
-    id: '$ow|p1',
-    regionId: ow,
-    displayName: 'P1',
-    ownerId: 'gp1',
-  );
-  final world = WorldState(
-    turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-    oldWorld: RegionData(provinces: [p1], units: const []),
-    newWorld: const RegionData(),
-    playerVisibilityByTile: const {},
-    playerProspectedTiles: const {},
-  );
-  const player = Player(id: 'gp1', displayName: 'Solo', isHuman: true);
-  return Game(
-    id: 'empty-diplo',
-    worldState: world,
-    players: const [player],
-    diplomacyRelations: const [],
-  );
-}
-
-/// Fixture for the discovery-via-visibility ACs (Refs #3341): the human GP
-/// `gp1` has fully-visible tile sight into a New-World province owned by Tribe
-/// `t1` but holds **no** `DiplomacyRelation` with the tribe. Per
-/// SPEC/ui/diplomacy-panel.md § Discovered factions, the panel must discover
-/// the tribe via `knownDiplomaticTargetFactionIds` and surface the default
-/// neutral first-contact standing.
-Game _gameWithTribeDiscoveredByVisibility() {
-  const nw = 'newWorld';
-  const ow = 'oldWorld';
-  final tribeProvince = Province(
-    id: '$nw|t1prov',
-    regionId: nw,
-    displayName: 'Tribe Land',
-    ownerId: 't1',
-  );
-  final homeProvince = Province(
-    id: '$ow|p1',
-    regionId: ow,
-    displayName: 'Home',
-    ownerId: 'gp1',
-  );
-  final world = WorldState(
-    turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 3),
-    oldWorld: RegionData(provinces: [homeProvince], units: const []),
-    newWorld: RegionData(provinces: [tribeProvince], units: const []),
-    playerVisibilityByTile: const {
-      'gp1': {'newWorld|t1prov|0|0': 'fullyVisible'},
-    },
-    playerProspectedTiles: const {},
-  );
-  const player = Player(id: 'gp1', displayName: 'Solo', isHuman: true);
-  return Game(
-    id: 'tribe-visibility',
-    worldState: world,
-    players: const [player],
-    tribes: const [Tribe(id: 't1', displayName: 'Tribe One')],
-    diplomacyRelations: const [],
-  );
-}
-
-/// Reads the percentage and tier span colors from the [RelativePowerLine]
-/// rendered inside the diplomacy row keyed by [factionId]. The line is a
-/// `Text.rich` whose root `TextSpan` children are
-/// `[prefix, percentage, separator, tier]`.
-({Color? pctColor, Color? tierColor}) _relativePowerSpanColors(
-  WidgetTester tester,
-  String factionId,
-) {
-  final lineFinder = find.descendant(
-    of: find.byKey(ValueKey('$kDiplomacyRowBodyKeyPrefix$factionId')),
-    matching: find.byType(RelativePowerLine),
-  );
-  final richText = tester.widget<RichText>(
-    find.descendant(of: lineFinder, matching: find.byType(RichText)).first,
-  );
-  // `Text.rich` nests the supplied root span under the effective-style span
-  // that `Text` builds, so the relative-power spans live one level deeper.
-  final root = (richText.text as TextSpan).children!.first as TextSpan;
-  final children = root.children!;
-  final pctSpan = children[1] as TextSpan;
-  final tierSpan = children[3] as TextSpan;
-  return (pctColor: pctSpan.style?.color, tierColor: tierSpan.style?.color);
-}
-
 void main() {
   suppressLogsForTests();
 
@@ -127,7 +39,7 @@ void main() {
     humanPlayerId = gameWithFactions.players.isNotEmpty
         ? gameWithFactions.players.first.id
         : 'gp1';
-    gameWithNoDiscovered = _gameWithNoDiscoveredFactions();
+    gameWithNoDiscovered = buildDiplomacyPanelGameWithNoDiscoveredFactions();
   });
 
   group('DiplomacyPanel', () {
@@ -222,24 +134,25 @@ void main() {
       },
     );
 
-    testWidgets('AC-6/AC-10: overture and FTP buttons shown disabled when invalid', (
-      WidgetTester tester,
-    ) async {
-      await bindDiplomacyTallTestSurface(tester);
-      await tester.pumpWidget(
-        buildDiplomacyPanel(
-          game: gameWithFactions,
-          humanPlayerId: humanPlayerId,
-          topology: topology,
-        ),
-      );
-      await pumpDiplomacyPanelBuilt(tester);
+    testWidgets(
+      'AC-6/AC-10: overture and FTP buttons shown disabled when invalid',
+      (WidgetTester tester) async {
+        await bindDiplomacyTallTestSurface(tester);
+        await tester.pumpWidget(
+          buildDiplomacyPanel(
+            game: gameWithFactions,
+            humanPlayerId: humanPlayerId,
+            topology: topology,
+          ),
+        );
+        await pumpDiplomacyPanelBuilt(tester);
 
-      expect(find.text('Consulate'), findsWidgets);
-      expect(find.text('Embassy'), findsWidgets);
-      expect(find.text('Establish FTP'), findsWidgets);
-      expect(find.text('Offer Peace'), findsWidgets);
-    });
+        expect(find.text('Consulate'), findsWidgets);
+        expect(find.text('Embassy'), findsWidgets);
+        expect(find.text('Establish FTP'), findsWidgets);
+        expect(find.text('Offer Peace'), findsWidgets);
+      },
+    );
 
     testWidgets('AC: Action buttons present for factions', (
       WidgetTester tester,
@@ -262,7 +175,6 @@ void main() {
         isTrue,
       );
     });
-
 
     testWidgets('AC: Pending orders show Cancel button, action button hidden', (
       WidgetTester tester,
@@ -305,7 +217,6 @@ void main() {
       expect(find.text('Cancel'), findsWidgets);
     });
 
-
     testWidgets(
       'AC-1: Empty state shows all three section headings + tribe placeholder',
       (WidgetTester tester) async {
@@ -336,7 +247,7 @@ void main() {
         await bindDiplomacyTallTestSurface(tester);
         await tester.pumpWidget(
           buildDiplomacyPanel(
-            game: _gameWithTribeDiscoveredByVisibility(),
+            game: buildDiplomacyPanelGameWithTribeDiscoveredByVisibility(),
             humanPlayerId: 'gp1',
             topology: const MapTopology(nodes: [], edges: []),
           ),
@@ -395,7 +306,9 @@ void main() {
           reason: 'Default "All" filter must render in --accent.',
         );
 
-        final gpOnlyButton = tester.widget<Text>(find.text('Great Powers only'));
+        final gpOnlyButton = tester.widget<Text>(
+          find.text('Great Powers only'),
+        );
         expect(
           gpOnlyButton.style?.color,
           EditorialMonoclePalette.muted,
@@ -438,7 +351,8 @@ void main() {
         expect(
           find.text('Minor Nations'),
           findsNothing,
-          reason: 'Minor Nations section must be hidden when GP-only is active.',
+          reason:
+              'Minor Nations section must be hidden when GP-only is active.',
         );
         expect(
           find.text('Tribes'),
@@ -446,9 +360,7 @@ void main() {
           reason: 'Tribes section must be hidden when GP-only is active.',
         );
 
-        final activeLabel = tester.widget<Text>(
-          find.text('Great Powers only'),
-        );
+        final activeLabel = tester.widget<Text>(find.text('Great Powers only'));
         expect(
           activeLabel.style?.color,
           EditorialMonoclePalette.accent,
@@ -479,7 +391,8 @@ void main() {
         expect(
           find.text('Great Powers'),
           findsNothing,
-          reason: 'Great Powers section must be hidden when Minors-only is active.',
+          reason:
+              'Great Powers section must be hidden when Minors-only is active.',
         );
         // Both Minor and Tribe sections may or may not appear depending on
         // discovered factions in the debug-init game; assert that whichever
@@ -563,7 +476,7 @@ void main() {
         return;
       }
       for (final r in stronger) {
-        final colors = _relativePowerSpanColors(tester, r.factionId);
+        final colors = relativePowerSpanColors(tester, r.factionId);
         expect(
           colors.pctColor,
           EditorialMonoclePalette.danger,
@@ -577,51 +490,53 @@ void main() {
       }
     });
 
-    testWidgets('AC: weaker/equal GP relative-power line renders in --success', (
-      WidgetTester tester,
-    ) async {
-      await bindDiplomacyTallTestSurface(tester);
-      await tester.pumpWidget(
-        buildDiplomacyPanel(
-          game: gameWithFactions,
-          humanPlayerId: humanPlayerId,
-          topology: topology,
-        ),
-      );
-      await pumpDiplomacyPanelBuilt(tester);
+    testWidgets(
+      'AC: weaker/equal GP relative-power line renders in --success',
+      (WidgetTester tester) async {
+        await bindDiplomacyTallTestSurface(tester);
+        await tester.pumpWidget(
+          buildDiplomacyPanel(
+            game: gameWithFactions,
+            humanPlayerId: humanPlayerId,
+            topology: topology,
+          ),
+        );
+        await pumpDiplomacyPanelBuilt(tester);
 
-      final rows = buildDiplomacyRows(
-        gameWithFactions,
-        topology,
-        humanPlayerId,
-        const Orders(),
-      );
-      final weakerOrEqual = rows
-          .where(
-            (r) =>
-                r.kind == FactionKind.greatPower &&
-                r.powerScore != null &&
-                r.playerPowerScore != null &&
-                powerComparisonPercent(r.powerScore!, r.playerPowerScore!) <= 0,
-          )
-          .toList();
-      if (weakerOrEqual.isEmpty) {
-        return;
-      }
-      for (final r in weakerOrEqual) {
-        final colors = _relativePowerSpanColors(tester, r.factionId);
-        expect(
-          colors.pctColor,
-          EditorialMonoclePalette.success,
-          reason: 'Weaker/equal GP percentage must use --success',
+        final rows = buildDiplomacyRows(
+          gameWithFactions,
+          topology,
+          humanPlayerId,
+          const Orders(),
         );
-        expect(
-          colors.tierColor,
-          EditorialMonoclePalette.success,
-          reason: 'Weaker/equal GP tier word must match the percentage color',
-        );
-      }
-    });
+        final weakerOrEqual = rows
+            .where(
+              (r) =>
+                  r.kind == FactionKind.greatPower &&
+                  r.powerScore != null &&
+                  r.playerPowerScore != null &&
+                  powerComparisonPercent(r.powerScore!, r.playerPowerScore!) <=
+                      0,
+            )
+            .toList();
+        if (weakerOrEqual.isEmpty) {
+          return;
+        }
+        for (final r in weakerOrEqual) {
+          final colors = relativePowerSpanColors(tester, r.factionId);
+          expect(
+            colors.pctColor,
+            EditorialMonoclePalette.success,
+            reason: 'Weaker/equal GP percentage must use --success',
+          );
+          expect(
+            colors.tierColor,
+            EditorialMonoclePalette.success,
+            reason: 'Weaker/equal GP tier word must match the percentage color',
+          );
+        }
+      },
+    );
 
     testWidgets(
       'AC: Great Power rows render the localized "Relative power:" prefix',
@@ -670,33 +585,32 @@ void main() {
   });
 
   group('DiplomacyPanel section headings (editorial-monocle)', () {
-    testWidgets(
-      'AC: Section heading text resolves to --accent color',
-      (WidgetTester tester) async {
-        await bindDiplomacyTallTestSurface(tester);
-        await tester.pumpWidget(
-          buildDiplomacyPanel(
-            game: gameWithFactions,
-            humanPlayerId: humanPlayerId,
-            topology: topology,
-          ),
-        );
-        await pumpDiplomacyPanelBuilt(tester);
+    testWidgets('AC: Section heading text resolves to --accent color', (
+      WidgetTester tester,
+    ) async {
+      await bindDiplomacyTallTestSurface(tester);
+      await tester.pumpWidget(
+        buildDiplomacyPanel(
+          game: gameWithFactions,
+          humanPlayerId: humanPlayerId,
+          topology: topology,
+        ),
+      );
+      await pumpDiplomacyPanelBuilt(tester);
 
-        final heading = tester.widget<Text>(find.text('Great Powers'));
-        expect(
-          heading.style?.color,
-          EditorialMonoclePalette.accent,
-          reason:
-              'Section heading must render in --accent per editorial-monocle.',
-        );
-        expect(
-          heading.style?.fontFamily,
-          'Cinzel',
-          reason: 'Section heading must use the editorial-monocle display font.',
-        );
-      },
-    );
+      final heading = tester.widget<Text>(find.text('Great Powers'));
+      expect(
+        heading.style?.color,
+        EditorialMonoclePalette.accent,
+        reason:
+            'Section heading must render in --accent per editorial-monocle.',
+      );
+      expect(
+        heading.style?.fontFamily,
+        'Cinzel',
+        reason: 'Section heading must use the editorial-monocle display font.',
+      );
+    });
 
     testWidgets(
       'AC: Section heading container exposes a 2 px --accent-dim bottom border',
@@ -774,51 +688,50 @@ void main() {
       },
     );
 
-    testWidgets(
-      'AC: Minor badge background --muted and foreground --bg-deep',
-      (WidgetTester tester) async {
-        await bindDiplomacyTallTestSurface(tester);
-        await tester.pumpWidget(
-          buildDiplomacyPanel(
-            game: gameWithFactions,
-            humanPlayerId: humanPlayerId,
-            topology: topology,
-          ),
-        );
-        await pumpDiplomacyPanelBuilt(tester);
+    testWidgets('AC: Minor badge background --muted and foreground --bg-deep', (
+      WidgetTester tester,
+    ) async {
+      await bindDiplomacyTallTestSurface(tester);
+      await tester.pumpWidget(
+        buildDiplomacyPanel(
+          game: gameWithFactions,
+          humanPlayerId: humanPlayerId,
+          topology: topology,
+        ),
+      );
+      await pumpDiplomacyPanelBuilt(tester);
 
-        final rows = buildDiplomacyRows(
-          gameWithFactions,
-          topology,
-          humanPlayerId,
-          const Orders(),
-        );
-        final hasMinor = rows.any((r) => r.kind == FactionKind.minor);
-        if (!hasMinor) return;
+      final rows = buildDiplomacyRows(
+        gameWithFactions,
+        topology,
+        humanPlayerId,
+        const Orders(),
+      );
+      final hasMinor = rows.any((r) => r.kind == FactionKind.minor);
+      if (!hasMinor) return;
 
-        final minorText = find.text('Minor').first;
-        final container = tester.widget<Container>(
-          find.ancestor(of: minorText, matching: find.byType(Container)).first,
-        );
-        final decoration = container.decoration as BoxDecoration;
-        expect(
-          decoration.color,
-          EditorialMonoclePalette.muted,
-          reason: 'Minor badge background must resolve to --muted.',
-        );
-        expect(
-          decoration.border,
-          isNull,
-          reason: 'Minor badge must not draw an outline border.',
-        );
-        final textWidget = tester.widget<Text>(minorText);
-        expect(
-          textWidget.style?.color,
-          EditorialMonoclePalette.bgDeep,
-          reason: 'Minor badge foreground must resolve to --bg-deep.',
-        );
-      },
-    );
+      final minorText = find.text('Minor').first;
+      final container = tester.widget<Container>(
+        find.ancestor(of: minorText, matching: find.byType(Container)).first,
+      );
+      final decoration = container.decoration as BoxDecoration;
+      expect(
+        decoration.color,
+        EditorialMonoclePalette.muted,
+        reason: 'Minor badge background must resolve to --muted.',
+      );
+      expect(
+        decoration.border,
+        isNull,
+        reason: 'Minor badge must not draw an outline border.',
+      );
+      final textWidget = tester.widget<Text>(minorText);
+      expect(
+        textWidget.style?.color,
+        EditorialMonoclePalette.bgDeep,
+        reason: 'Minor badge foreground must resolve to --bg-deep.',
+      );
+    });
 
     testWidgets(
       'AC: Tribe badge outlined with --muted border, transparent background',
@@ -839,8 +752,9 @@ void main() {
           humanPlayerId,
           const Orders(),
         );
-        final tribeRows =
-            rows.where((r) => r.kind == FactionKind.tribe).toList();
+        final tribeRows = rows
+            .where((r) => r.kind == FactionKind.tribe)
+            .toList();
         if (tribeRows.isEmpty) return;
         final tribeRow = tribeRows.first;
 
@@ -857,9 +771,7 @@ void main() {
         );
         await tester.pump();
         final container = tester.widget<Container>(
-          find
-              .ancestor(of: tribeText, matching: find.byType(Container))
-              .first,
+          find.ancestor(of: tribeText, matching: find.byType(Container)).first,
         );
         final decoration = container.decoration as BoxDecoration;
         expect(
@@ -902,12 +814,21 @@ void main() {
           final widget = tester.widget<Text>(finder.first);
           final color = widget.style?.color;
           // Reject the prior hardcoded Material palette for these labels.
-          expect(color, isNot(equals(Colors.blue)),
-              reason: '$label badge must not use Colors.blue.');
-          expect(color, isNot(equals(Colors.grey)),
-              reason: '$label badge must not use Colors.grey.');
-          expect(color, isNot(equals(Colors.orange)),
-              reason: '$label badge must not use Colors.orange.');
+          expect(
+            color,
+            isNot(equals(Colors.blue)),
+            reason: '$label badge must not use Colors.blue.',
+          );
+          expect(
+            color,
+            isNot(equals(Colors.grey)),
+            reason: '$label badge must not use Colors.grey.',
+          );
+          expect(
+            color,
+            isNot(equals(Colors.orange)),
+            reason: '$label badge must not use Colors.orange.',
+          );
         }
       },
     );
